+++ Plugins/SQLImportPlugin/SQLImportOptions.swift
@@ -4,8 +4,10 @@
 //
 
 import Foundation
+import TableProPluginKit
 
 struct SQLImportOptions: Equatable, Codable {
+    var errorHandling: ImportErrorHandling = .stopAndRollback
     var wrapInTransaction: Bool = true
     var disableForeignKeyChecks: Bool = true
 }